server-side caching implemented (#11)

+++ src/services/postgres/NotesService.js
@@ -7,9 +7,10 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const { mapDBToModel } = require('../../utils/mapDBToModel');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this.collaborationService = collaborationService;
+    this.cacheService = cacheService;
   }
 
   async addNote({
@@ -29,21 +30,29 @@ class NotesService {
     if (!result.rows[0].id) {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
-
+    await this.cacheService.delete(`notes:${owner}`);
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-      LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner = $1 OR collaborations.user_id = $1
-      GROUP BY notes.id`,
-      values: [owner],
-    };
+    try {
+      const result = await this.cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      await this.cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -66,7 +75,7 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title=$2, body=$3, tags=$4 , updated_at=$5 where id = $1 RETURNING id',
+      text: 'UPDATE notes SET title=$2, body=$3, tags=$4 , updated_at=$5 where id = $1 RETURNING *',
       values: [id, title, body, tags, updatedAt],
     };
 
@@ -75,11 +84,15 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+    const { owner } = result.rows[0];
+    await this.cacheService.delete(`notes:${owner}`);
+
+    return result.rows.map((note) => ({ title: note.title, tags: note.tags, body: note.body }))[0];
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes where id = $1 RETURNING id',
+      text: 'DELETE FROM notes where id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -88,6 +101,9 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    await this.cacheService.delete(`notes:${owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
